Added MathParse method for strings
string.MathParse takes a string, processes it as a mathematical formula (optionally with a dictionary of variables), and returns the result.

This function is currently only here as a proof of concept and is not fit for common use as-is. I do have plans to rework it to execute faster/split parsing and executing into separate methods, but can't do those right now.